book-backend/pyproject.toml
[project]
name = "book-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "python-dotenv>=1.2.1",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "PyJWT>=2.10.1",
    "openai-agents>=0.6.2",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "qdrant-client>=1.8.0",
    "sentence-transformers>=2.2.0",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "ebooklib>=0.18",
    "speechrecognition>=3.10.0",
    "pydub>=0.25.1",
    "openai-whisper>=20231117",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-benchmark>=4.0.0",
    "locust>=2.17.0",
]
